requirements.txt
# Core dependencies for PocketFlow examples
anthropic>=0.18.0
python-dotenv>=1.0.0

# Optional but recommended
requests>=2.31.0  # For web tools
numpy>=1.24.0     # For calculations
pandas>=2.0.0     # For data processing

# Development tools (optional)
pytest>=7.0.0     # For testing
black>=23.0.0     # For formatting
flake8>=6.0.0     # For linting

# PocketFlow itself (if not using local development)
# pocketflow>=0.0.2